np types

+++ score_splitter.py
@@ -106,7 +106,7 @@ class Score:
         horiz_sum_verts = verts_norm.sum(axis=1)
         indices = find_peaks(-1*horiz_sum_verts)[0]
         minima = horiz_sum_verts[indices]
-        horiz_sum_hist = np.bincount(horiz_sum_verts)
+        horiz_sum_hist = np.bincount(horiz_sum_verts.astype(int))
         avg_min = np.argmax(horiz_sum_hist)
         # the most important part
         # if we take the max, it starts to kick out values we don't want it to
